+++ tokio-postgres/src/connect_raw.rs
@@ -13,6 +13,7 @@ use postgres_protocol::authentication::sasl;
 use postgres_protocol::authentication::sasl::ScramSha256;
 use postgres_protocol::message::backend::{AuthenticationSaslBody, Message};
 use postgres_protocol::message::frontend;
+use std::borrow::Cow;
 use std::collections::{HashMap, VecDeque};
 use std::io;
 use std::pin::Pin;
@@ -96,8 +97,13 @@ where
         delayed: VecDeque::new(),
     };
 
-    startup(&mut stream, config).await?;
-    authenticate(&mut stream, config).await?;
+    let user = config
+        .user
+        .as_deref()
+        .map_or_else(|| Cow::Owned(whoami::username()), Cow::Borrowed);
+
+    startup(&mut stream, config, &user).await?;
+    authenticate(&mut stream, config, &user).await?;
     let (process_id, secret_key, parameters) = read_info(&mut stream).await?;
 
     let (sender, receiver) = mpsc::unbounded();
@@ -107,13 +113,17 @@ where
     Ok((client, connection))
 }
 
-async fn startup<S, T>(stream: &mut StartupStream<S, T>, config: &Config) -> Result<(), Error>
+async fn startup<S, T>(
+    stream: &mut StartupStream<S, T>,
+    config: &Config,
+    user: &str,
+) -> Result<(), Error>
 where
     S: AsyncRead + AsyncWrite + Unpin,
     T: AsyncRead + AsyncWrite + Unpin,
 {
     let mut params = vec![("client_encoding", "UTF8")];
-    params.push(("user", &config.user));
+    params.push(("user", user));
     if let Some(dbname) = &config.dbname {
         params.push(("database", &**dbname));
     }
@@ -139,7 +149,11 @@ where
         .map_err(Error::io)
 }
 
-async fn authenticate<S, T>(stream: &mut StartupStream<S, T>, config: &Config) -> Result<(), Error>
+async fn authenticate<S, T>(
+    stream: &mut StartupStream<S, T>,
+    config: &Config,
+    user: &str,
+) -> Result<(), Error>
 where
     S: AsyncRead + AsyncWrite + Unpin,
     T: TlsStream + Unpin,
@@ -162,7 +176,6 @@ where
         Some(Message::AuthenticationMd5Password(body)) => {
             can_skip_channel_binding(config)?;
 
-            let user = &config.user;
             let pass = config
                 .password
                 .as_ref()