+++ tokio-postgres/CHANGELOG.md
@@ -1,6 +1,37 @@
 # Change Log
 
-## v0.7.9
+## Unreleased
+
+## v0.7.11 - 2024-07-21
+
+### Fixed
+
+* Fixed handling of non-UTF8 error fields which can be sent after failed handshakes.
+* Fixed cancellation handling of `TransactionBuilder::start` futures.
+
+### Added
+
+* Added `table_oid` and `field_id` fields to `Columns` struct of prepared statements.
+* Added `GenericClient::simple_query`.
+* Added `#[track_caller]` to `Row::get` and `SimpleQueryRow::get`.
+* Added `TargetSessionAttrs::ReadOnly`.
+* Added `Debug` implementation for `Statement`.
+* Added `Clone` implementation for `Row`.
+* Added `SimpleQueryMessage::RowDescription`.
+* Added `{Client, Transaction, GenericClient}::query_typed`.
+
+### Changed
+
+* Disable `rustc-serialize` compatibility of `eui48-1` dependency
+* Config setters now take `impl Into<String>`.
+
+## v0.7.10 - 2023-08-25
+
+## Fixed
+
+* Defered default username lookup to avoid regressing `Config` behavior.
+
+## v0.7.9 - 2023-08-19
 
 ## Fixed
 
@@ -13,7 +44,7 @@
 * Added support for the `load_balance_hosts` config option to randomize connection ordering.
 * The `user` config option now defaults to the executing process's user.
 
-## v0.7.8
+## v0.7.8 - 2023-05-27
 
 ## Added